# Ensure Context is noncopyable