--- a/src/binding.cpp
+++ b/src/binding.cpp
@@ -1093,7 +1093,8 @@ struct Context {
     uint32_t mask;
     double relev;
 
-    Context(Context const& c) = default;
+    Context(Context const& c) = delete;
+    Context& operator=(Context const& c) = delete;
     Context(Cover && cov,
             uint32_t mask,
             double relev)