--- a/src/binding.cpp
+++ b/src/binding.cpp
@@ -24,14 +24,14 @@ Nan::Persistent<FunctionTemplate> RocksDBCache::constructor;
 rocksdb::Status OpenDB(const rocksdb::Options& options, const std::string& name, std::unique_ptr<rocksdb::DB>& dbptr) {
     rocksdb::DB* db;
     rocksdb::Status status = rocksdb::DB::Open(options, name, &db);
-    dbptr = std::move(std::unique_ptr<rocksdb::DB>(db));
+    dbptr = std::unique_ptr<rocksdb::DB>(db);
     return status;
 }
 
 rocksdb::Status OpenForReadOnlyDB(const rocksdb::Options& options, const std::string& name, std::unique_ptr<rocksdb::DB>& dbptr) {
     rocksdb::DB* db;
     rocksdb::Status status = rocksdb::DB::OpenForReadOnly(options, name, &db);
-    dbptr = std::move(std::unique_ptr<rocksdb::DB>(db));
+    dbptr = std::unique_ptr<rocksdb::DB>(db);
     return status;
 }