This fixes an odd case (still need to research more into the why here) where Context objects were still being copied rather than moved (even though they had a move constructor enabled).

By moving these large objects we should avoid many allocations and help speed up performance. 

On OS X this leads to a significant speedup. On Linux the speedup is less but still noticable.

OSX shows:

#### Master

```
# coalesceMulti
ok 117 coalesceMulti @ 7.56ms
# coalesceMulti proximity
ok 118 coalesceMulti + proximity @ 7.44ms
```

#### This branch

```
ok 117 coalesceMulti @ 6.62ms
# coalesceMulti proximity
ok 118 coalesceMulti + proximity @ 6.46ms
```