@springmeyer 2017-04-05
Latest results with `std::partial_sort`:

```
TAP version 13
# coalesceSingle
ok 1 coalesceSingle @ 0.42ms
# coalesceSingle proximity
ok 2 coalesceSingle + proximity @ 0.52ms
# coalesceMulti
ok 3 coalesceMulti @ 4.84ms
# coalesceMulti proximity
ok 4 coalesceMulti + proximity @ 4.78ms
```