@springmeyer 2017-04-05
Because the partial sort speeds things up so much I took a look at profiling results with 45c91ed and after partial sort (83b555d) on OS X.

In both cases, for one thread sampled, 22% of the thread is busy in `coalesceMulti`. But the speedup with partial sort is obvious because it moves the sort from being where the largest time taken to only 4th most expensive.

Before:
  - 7.2% is taken in `std::sort` of `ContextSortByRelev`
  - 5.3% is taken in deallocation of memory (`free`)
  - 3.8% is taken in allocation of memory (`malloc/new`)
  - 2.2% is taken in the `std::sort` in `__getmatching`
  - Everything else is less than 1%

After:
  - 5.9% is taken in deallocation of memory (`free`)
  - 5.1% is taken in allocation of memory (`malloc/new`)
  - 3.8% is taken in the `std::sort` in `__getmatching`
  - 1.3% is taken in `std::partial_sort` of `ContextSortByRelev`
  - Everything else is less than 1%