@apendleton 2017-04-05
Sidenote as far as next steps here: `__getmatching` is actually overloaded and has two implementations, one for rocksdb and one for the in-memory cache. We only actually use the former in production, and only the latter calls `std::sort` (the rocksdb version sorts lazily). We may want to adjust these benchmarks accordingly as I wonder if the performance characteristics will be different... I expect that the rocksdb implementation may well be slower because it doesn't hold everything in ram, though, so it might make benchmarks a little noisier.