@springmeyer 2017-04-05
Thanks @apendleton - per chat I've love help on adjusting the benchmark to test the right thing. Feel free to commit to this branch or another with any fixes there.

The other next steps here I see are:

  - Figure out if the `partial_sort` can be done on a fixed, predictable `N`. Currently we depend on filtering to a max of `40` after sorting and so the `partial_sort` needs to run on more than `40` values. If we could dedupe features before sorting that would allow the sort to be much faster.
  - Test in carmen and see if the perf gains above also translate to production before spending too much more time here.