@springmeyer 2017-04-06
Noting that while OS X results are impressive, results on linux/travis are not. So I need to see if travis is lying (very possible) or perhaps libstdc++ (defalt standard library on linux) is not as optimized for `partial_sort` as libc++ (OSX):

### master:

```
# coalesceSingle
ok 115 coalesceSingle @ 0.48ms
# coalesceSingle proximity
ok 116 coalesceSingle + proximity @ 0.68ms
# coalesceMulti
ok 117 coalesceMulti @ 7.56ms
# coalesceMulti proximity
ok 118 coalesceMulti + proximity @ 7.44ms
```

### 45c91ed (this branch, before partial_sort)

```
# coalesceSingle
ok 115 coalesceSingle @ 0.5ms
# coalesceSingle proximity
ok 116 coalesceSingle + proximity @ 0.68ms
# coalesceMulti
ok 117 coalesceMulti @ 7.22ms
# coalesceMulti proximity
ok 118 coalesceMulti + proximity @ 7.52ms
```


### 83b555d (this branch, after partial_sort)

```
# coalesceSingle
ok 115 coalesceSingle @ 0.82ms
# coalesceSingle proximity
ok 116 coalesceSingle + proximity @ 0.84ms
# coalesceMulti
ok 117 coalesceMulti @ 5.68ms
# coalesceMulti proximity
ok 118 coalesceMulti + proximity @ 7.44ms
```