@mourner 2017-04-06
> If we could dedupe features before sorting that would allow the sort to be much faster.

Yes! Also note that for deduping, the existing code in `coalesceFinalize` uses `std::map`. Since we're deduping by a uint `id`, using `std::unordered_set` should be much faster since insertion/lookup is constant time on average.