@springmeyer 2017-04-28
Update on my thinking on this PR:

  - The non-copyable fixes are 👍. While they may not show huge perf gains in local testing they should reduce allocations and therefore should help most under high load
  - The `partial_sort` approach feels very promising and getting close but not quite right. The next actions at https://github.com/mapbox/carmen-cache/pull/93#issuecomment-292028515 still stand. Ideally the partial sort could be done on a fixed N. So, we should - as @mourner - mentions revisit how deduping works.

I'm going to pause here since `revisit how deduping works` is above my ability to focus on the code right now. I would be happy to spirit guide someone who wants to pick this up.