@springmeyer 2017-05-23
I've removed the experimental/not-quite-yet-ideal `partial_sort` code from this PR in 57ec8fc12e3491df6f767397b945a727474e916a.

This means that I think this is ready to be tested in production since the remaining changes are clear improvements.

 - Avoiding double move of `std::unique_ptr<rocksdb::DB>` (and a compiler warning on pessimizing move)
 - Deleting the automatic-compiler-generated constructor, copy constructor, and assignment constructor of the `Context` class. This ensures that the `Context` is truly "noncopyable" and therefore only movable. This will save on memory allocations on linux where otherwise the object was being copied rather than moved.
 - Avoiding reserving too much memory when calling `context.reserve` and there were less than 40 contexts to pick from.

/cc @KaiBot3000 @aarthykc - would you be interested in running this out? My hope is that this will help reduce memory usage in production and therefore might also help performance. The steps would be:

 - [ ] Tweak the `version` in package.json to something unique for testing like `0.18.0-moveonly`
 - [ ] Publish C++ binaries: `git commit -am "[publish binary]"
 - [ ] Test in carmen and if things look good
 - [ ] Stage, and if things look good
 - [ ]  Run out an official release of carmen-cache by incrementing the version to `0.19.0` and then:
   - [ ] Publishing binaries again
   - [ ] Then running `npm publish`
   - [ ] Updating `carmen` to `0.19.0` of carmen cache

If #96 is looking good, this could also be merged into master and tested alongside that as it is deployed.