@springmeyer 2018-02-23
Note, after #116 lands I think it would be a good time to revisit this and get it landed. Basically making sure that the `Context` is noncopyable to ensure it is moved by the compiler in all cases:

```diff
+    Context() = delete;
+    Context(Context const& c) = delete;
+    Context& operator=(Context const& c) = delete;
```

Also as a future idea/ticket: would be great to apply learnings on sorting optimizations from vtquery. Whoever wants to pick this up reach out to @mapsam for tips on sorting optimizations in C++.