@apendleton 2018-03-13
We made Context noncopyable in https://github.com/mapbox/carmen-cache/pull/116 . @springmeyer closing as I think ultimately that's all we wanted to carry over from here, but if there was other stuff from the earlier iterations of this PR that you wanted us to do as well, feel free to reopen.