@springmeyer 2018-03-13
Sounds good @apendleton - that should help. I've ticketed https://github.com/mapbox/carmen-cache/issues/120 to log the idea of using `partial_sort` in the future.